refactor: prepare for narrative-ui integration
