+++ src/components/Game/Game.jsx
@@ -1,12 +1,12 @@
 import { Canvas } from "@react-three/fiber"
 import classes from "./Game.module.scss"
-import { Introduction } from "../scenes/Introduction/Introduction"
+import { RegionEntranceOne } from "../scenes/RegionEntranceOne/RegionEntranceOne"
 
 export const Game = () => {
   return (
     <div className={classes.container}>
       <Canvas>
-        <Introduction />
+        <RegionEntranceOne />
       </Canvas>
     </div>
   )
